Switched to JUnit 4.13 from Junit.Jupiter

See merge request se-2021/group-58/ezshop!5

+++ src/test/java/it/polito/ezshop/data/AcceptableIssueOrder.java
@@ -4,7 +4,7 @@ package it.polito.ezshop.data;
 import it.polito.ezshop.exceptions.*;
 import org.junit.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.Assert.*;
 
 public class AcceptableIssueOrder {
 
@@ -36,7 +36,7 @@ public class AcceptableIssueOrder {
     {
         EZShop shop = new EZShop();
         shop.login("admin","ciao");
-        assertEquals(-1,shop.issueOrder("652343546457",5,30.0));
+        assertEquals(-1,shop.issueOrder("652343546457",5,30.0).intValue());
     }
 
     @Test
